requirements.txt
Flask==2.3.2
google-cloud-speech==2.15.0
google-cloud-texttospeech==2.11.0
Werkzeug==2.3.7